refactor: replacing wrapper in index with actual nerdlet (fixes #5)
